temp: Remove LOS until it can be fully fixed
I use offsetx and offsety in sprite_hero to prevent true_points
(the point on the actual corner) and the raycast points (the point
past the corner) from being in the same line.  If they were
at the same line to the hero, they couldn't be sorted by angle
and the LOS triangles wouldn't render in the right order.
But due to the offset, in some circumstances, the poly doesn't
draw correctly.
Turning off for now, going to focus on AI instead for better
gameplay.